backend/requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cmake==3.18.2.post1
cryptography==3.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
dlib==19.21.0
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
libcst==0.3.10
mypy-extensions==0.4.3
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
Pillow==7.2.0
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
